transcribe_cosmology additional calculations
We're calculating the fractional densities since Brenda uses those. We're also adding the omega_nu density to the omega_cdm density since Brenda doesn't distinguish these....

But maybe this is wrong!? I'm pretty sure it's correct because we're doing this trick as part of the effort to get a sigma12 value. What sigma12 value do we care about? The sigma12 of the MEMNeC, which by definition lumps the neuts into the CDM