Clean up scale_sigma12 documentation
Also in this commit: we're using the sigma12 emu to normalize the sigma12 input vector.

Since we were able to implement this normalization without any explicit reference to the priors, we'll be able to delete the prior_file_to_array fn. assuming all goes well with the other two normalization steps.