Repair format error in fill_in_defaults
I mistakenly assumed that we already had a Brenda Cosmology by that point. The fn has been rewritten to refer to a cosmo dictionary.

Additionally, we fixed a couple of misspellings of the key "As" and added a poor docstring to error_check_cosmology.

Lastly, we've updated the installer script. However, we won't test installing this package for a while yet.