De-nest LGF
This concludes the ragged array saga--we've really fixed the problem now, it seems.

Also in this commit:
* Updated error_check_cosmology with an excerpt from transcribe_cosmology.

We've completed the first error-free run, but we still have to make sure that the numbers are reasonable. Furthermore, there are other cases that we need to test, such as when sigma12 is already specified.