cosmology_to_Pk docstring, part 1
I want to convert the parameter descriptions to the reST format, but there are a lot of parameters so I'm breaking this into steps.

We've pulled most of this docstring from the function transcribe_cosmology, which now serves a slightly different purpose. We'll come back later and add a new docstring from transcribe_cosmology.

Also in this commit: add important data files to the package_data field of the setup.py script. Adjust .gitignore to acknowledge this as important.