Changed my mind about 300k
It's inconsistent to include one data file and not all of them. But if I were to include all of them, I would have to include 2.5 GB worth of emulators. That would be especially bad considering I am planning to replace these emulators in the next week, so users cloning the repo would be forced to clone all of the old emulators too...

Also in this commit:
* Further greatly simplified the gitignore, which was bloated with LaTeX stuff which is no longer relevant since this, unlike the parent repo, is a pure-code repo.