Correct logic in scale_sigma12
I'm pretty sure that the function up to this point has been incorrect, although I'm saying this after looking at the code, rather than after looking at any particular results.

The problem was that we were using the DEFAULT_LGF as our baseline for computing the growth_ratio, but we should have been using the LGF of the cosmology which we inserted into the sigma12 emulator.

This cosmology was similar to the default cosmology (renamed to DEFAULT_COSMO_DICT for greater clarity), but importantly changed the values of omega_b and omega_cdm, which impact the value of the LGF.

Since this baseline cosmology is different every time, we've needed to add a new parameter to the scale_sigma12 fn. The fn add_sigma12 has been rewritten to compute this parameter, and we've added internal documentation to explain the calculation.