Rewrite roshambo section of transcribe
I'm not sure I've made the situation any better, but I think it's a little clearer now. It's become obvious to me that there are more computationally efficient ways of writing this section, but they would be even harder to read than this approach is. I'm not sure if I want to rewrite it to be even less efficient, because I'm still concerned that readers will get lost here.