Completely ditch kwargs
However, we've learned that ** is simply an operator, so we could theoretically bring back kwargs whenever we want. That may even make it easier to write docstrings, because I'll be able to describe cosmological parameters as function parameters, rather than as possible entries in an input dictionary.

Also in this commit:
* Add two new very important fns: get_Pk_interpolator and get_Delta2k_interpolator.
* Added test_cases.py, which should help us to procure the timings we'll need for the paper. Of course, we're going to want to:
A. Look up on the internet whether this is the best way to time functions.
B. Run this script many times on many cosmologies, to get timing statistics