Critical bug fix
We were recording the omega_DE value under the key "omega_DE," but we should have recorded it under "omega_de." This also disrupted the value of "Omega_DE," so no wonder the sigma12 values were wrong. Things look good for model 0 but we still need to test for a large number of random cosmologies.

This could be the end of the sigma12 bugs, but we still have to resolve the emulator usage bug...