Fix incomplete loading of cosmologies
We were being too hasty, and consequently the loaded cosmologies were missing critical parameters for proper sigma12 estimation.

Also in this commit:
* Fix a typo in ei that caused a crash
* Fix the lack of redshift in generating the solution key.

At this point, the percent errors in the test set have been reduced significantly, but it seems like we've still got a long way to go. The first two percent errors that we see, for example, are 6 and 17. We might be able to get away with 0.1% error, but that's already pushing it since we're just talking about error on sigma12, not even P(k) yet.