One of two estimate_sigma12 errors fixed
This is the critical one: we forgot to recalculate omega_m, Omega_m, omega_de, and Omega_DE when creating the emu_cosmology. This meant that our LGFs were seriously messed up!

The other bug is that the A_s values don't agree exactly between two ways of generating the default brenda cosmology. Let's address this and then we'll be ready for a new batch of sigma12 values!

I can feel it--we're really close to the end of the sigma12 saga. Then we can start debugging the emulators.